# TrackIt.Frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrackIt API",
    "version": "v1"
  },
  "paths": {
...
    "/api/budgets": {
      "post": {
        "tags": [
          "Budget"
        ],
        "summary": "Создание лимита",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetCommand"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Budget"
        ],
        "summary": "Получение всех лимитов",
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "description": "Идентификатор категории",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Предел",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetDtoPagedList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetDtoPagedList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetDtoPagedList"
                }
              }
            }
          }
        }
      }
    },
    "/api/budgets/{id}": {
      "put": {
        "tags": [
          "Budget"
        ],
        "summary": "Обновление лимита",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Budget"
        ],
        "summary": "Удаление лимита",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Budget"
        ],
        "summary": "Получение лимита по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedBudgetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedBudgetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedBudgetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories": {
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Создание категории",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryCommand"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Получение всех категорий",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Предел",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDtoPagedList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDtoPagedList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDtoPagedList"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Обновление категории",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Удаление категории",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Получение категории по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedCategoryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedCategoryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedCategoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments": {
      "post": {
        "tags": [
          "PlannedPayment"
        ],
        "summary": "Создание платежа",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedPaymentCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedPaymentCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedPaymentCommand"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedPaymentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedPaymentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedPaymentDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "PlannedPayment"
        ],
        "summary": "Получение всех платежей",
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "description": "Идентификатор категории",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Предел",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedPaymentDtoPagedList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedPaymentDtoPagedList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedPaymentDtoPagedList"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/{id}": {
      "put": {
        "tags": [
          "PlannedPayment"
        ],
        "summary": "Обновление платежа",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedPaymentCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedPaymentCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedPaymentCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "PlannedPayment"
        ],
        "summary": "Удаление платежа",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "PlannedPayment"
        ],
        "summary": "Получение платежа по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedPlannedPaymentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedPlannedPaymentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedPlannedPaymentDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Величина лимита",
            "format": "double"
          },
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель лимита бюджета"
      },
      "BudgetDtoPagedList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetDto"
            },
            "description": "Элементы",
            "nullable": true
          },
          "total": {
            "type": "integer",
            "description": "Всего элементов",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Пагинированный список"
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор категории",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название категории",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CategoryType"
          },
          "user_id": {
            "type": "string",
            "description": "Идентификатор пользователя",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Модель категории"
      },
      "CategoryDtoPagedList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            },
            "description": "Элементы",
            "nullable": true
          },
          "total": {
            "type": "integer",
            "description": "Всего элементов",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Пагинированный список"
      },
...
      "CategoryType": {
        "enum": [
          1,
          -1
        ],
        "type": "integer",
        "description": "Тип категории",
        "format": "int32"
      },
      "CreateBudgetCommand": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Величина лимита",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "Запрос на создание запланированного платежа"
      },
      "CreateCategoryCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название категории",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CategoryType"
          }
        },
        "additionalProperties": false,
        "description": "Запрос на создание категории"
      },
      "CreatePlannedPaymentCommand": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Объем оплаты",
            "format": "double"
          },
          "due_date": {
            "type": "string",
            "description": "Дата",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "description": "Описание",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Запрос на создание запланированного платежа"
      },
...
      "DetailedBudgetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Величина лимита",
            "format": "double"
          },
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "description": "Дата обновления",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Детализированная модель лимита бюджета"
      },
      "DetailedCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор категории",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название категории",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CategoryType"
          },
          "user_id": {
            "type": "string",
            "description": "Идентификатор пользователя",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "description": "Дата обновления",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Детализированная модель категории"
      },
      "DetailedPlannedPaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Объем оплаты",
            "format": "double"
          },
          "due_date": {
            "type": "string",
            "description": "Дата",
            "format": "date-time"
          },
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Описание",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "description": "Дата обновления",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Детализированная модель запланированной платы"
      },
...
      "PlannedPaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Объем оплаты",
            "format": "double"
          },
          "due_date": {
            "type": "string",
            "description": "Дата",
            "format": "date-time"
          },
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель запланированной платы"
      },
      "PlannedPaymentDtoPagedList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlannedPaymentDto"
            },
            "description": "Элементы",
            "nullable": true
          },
          "total": {
            "type": "integer",
            "description": "Всего элементов",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Пагинированный список"
      },
...
      "UpdateBudgetCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор запланированного платежа",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Величина лимита",
            "format": "double"
          },
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Запрос на обновление запланированного платежа"
      },
      "UpdateCategoryCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор категории",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название категории",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CategoryType"
          }
        },
        "additionalProperties": false,
        "description": "Запрос на обновление категории"
      },
      "UpdatePlannedPaymentCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор запланированного платежа",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Объем оплаты",
            "format": "double"
          },
          "due_date": {
            "type": "string",
            "description": "Дата",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "description": "Описание",
            "nullable": true
          },
          "category_id": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Запрос на обновление запланированного платежа"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Введите ваш JWT-токен в формате: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}